rename and add constants
