+++ src/cliAction/startStopDeleteProject/handler_runCmdType.go
@@ -1,6 +1,10 @@
 package startStopDeleteProject
 
-import "context"
+import (
+	"context"
+
+	"github.com/zerops-io/zcli/src/constants"
+)
 
 func (h *Handler) Run(ctx context.Context, config RunConfig, actionType string) error {
 	projectId, err := h.getProjectId(ctx, config)
@@ -8,9 +12,9 @@ func (h *Handler) Run(ctx context.Context, config RunConfig, actionType string) 
 		return err
 	}
 
-	if actionType == "start" {
+	if actionType == constants.Start {
 		return h.RunStart(ctx, config, projectId)
-	} else if actionType == "stop" {
+	} else if actionType == constants.Stop {
 		return h.RunStop(ctx, config, projectId)
 	} else {
 		return h.RunDelete(ctx, config, projectId, actionType)
