+++ src/constants/actionTypes.go
@@ -1,0 +1,9 @@
+package constants
+
+const (
+	Start  = "start"
+	Stop   = "stop"
+	Delete = "delete"
+	// confirmed delete (cmd with --confirm)
+	YDelete = "y-delete"
+)
